- Point Cloud visualizer now stores labels with points.  - removed a bunch of outdated todos  - enable switching between test set and training set projection view  - fixed a bunch of warnings about casts from int to unsigned char.  - RankResults.fsx sorts by mean training error now.

+++ LvqEmn/LvqLib/SmartSum.h
@@ -10,7 +10,7 @@ class SmartSum {
 	ArrayT meanX, sX;
 	double weightSum;
 public:
-	SmartSum(int dims) : meanX(ArrayT::Zero(dims) ), sX(ArrayT::Zero(dims)), weightSum(0.0) {}
+	SmartSum(size_t dims) : meanX(ArrayT::Zero(dims) ), sX(ArrayT::Zero(dims)), weightSum(0.0) {}
 	SmartSum() : meanX(ArrayT::Zero(CompileTimeDims) ), sX(ArrayT::Zero(CompileTimeDims)), weightSum(0.0) {
 		assert(CompileTimeDims != Eigen::Dynamic && CompileTimeDims>0);
 	}
